% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_country_codes.R
\name{country_regex_to_iso3c}
\alias{country_regex_to_iso3c}
\title{Country regex to iso3c}
\usage{
country_regex_to_iso3c(country_name)
}
\arguments{
\item{country_name}{a string containing a country name}
}
\value{
a string containing the three letter iso3c country code
}
\description{
Convert a country name to iso3c country code using regular expressions.
}
\details{
This function uses regular expressions to convert a country name written in
a variety of formats into iso3c code.  For example, the function will convert both "Russia" and "Russian Federation"
to the correct iso3c code "RUS".  This can be particularly useful when merging data from multiple sources.
From iso3c country names can easily be converted into other formats. This is a simple convenience function written
on top of the \code{countrycode} package.
}
\examples{
country_regex_to_iso3c("Russia")
country_regex_to_iso3c("Russian Federation")
}
