% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_country_codes.R
\name{iso2c_to_iso3c}
\alias{iso2c_to_iso3c}
\title{iso2c to iso3c}
\usage{
iso2c_to_iso3c(iso2c)
}
\arguments{
\item{iso2c}{a string containing an iso2c country code}
}
\value{
string containing an iso3c country code
}
\description{
Convert an iso2c country code into an iso3c country code
}
\details{
This function will convert an iso2c country code into an iso3c country code.
It is a simple convenience function built on top of the \code{countrycode} package.
While somewhat arbitrary, iso3c may be a better default because the iso2c for
Namibia is "NA", which can cause innadvertant (and often silent) errors if R interprets it as
an \code{NA} value.   Better to go with the iso3c code where Namibia is "NAM" and avoid those problems.
}
\examples{
iso2c_to_iso3c("NA")

}
