% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_country_codes.R
\name{iso3c_to_country_name}
\alias{iso3c_to_country_name}
\title{iso3c to country name}
\usage{
iso3c_to_country_name(iso3c)
}
\arguments{
\item{iso3c}{a string containing an iso3c country code}
}
\value{
a string containing the country name
}
\description{
Convert an iso3c country code into a country name
}
\details{
This function converts an iso3c country code into the country name.  It is simple convenience function written on top of
the \code{countrycode package}. When working with data coming from different sources, a common workflow may be to use \code{country_regex_to_iso3c}
to convert country names written in different ways ("Ivory Coast" vs. "Cote D'Ivoire") into a unique iso3c country code, and then to use
\code{iso3c_to_country_name} to convert the iso3c codes back into uniform country names.
}
\examples{
iso3c_to_country_name("NAM")

"Bolivarian Republic of Venezuela" \%>\%
  country_regex_to_iso3c() \%>\%
  iso3c_to_country_name()


}
