% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_colors.R
\docType{data}
\name{wb_colors}
\alias{wb_colors}
\title{World Bank Colors}
\format{
A character vector of color names and hex values
}
\source{
Colors come from the World Bank's official \href{http://pubdocs.worldbank.org/en/546361452012867685/wbg-branding-visual-identity-guideline.pdf}{branding and visual identity guidelines} color names come \href{http://chir.ag/projects/name-that-color/#6195ED}{this website}. I didn't make them up.  I swear.
}
\usage{
wb_colors
}
\description{
The list of colors from the World Bank's official branding and visual identity guidelines.
}
\keyword{datasets}
