% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_color_palette_functions.R
\name{wb_cols}
\alias{wb_cols}
\title{Return function to interpolate a drsimonj color palette}
\usage{
wb_cols(...)
}
\arguments{
\item{...}{Optional character color names}
}
\description{
Return function to interpolate a drsimonj color palette
}
