% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.r
\docType{data}
\name{wb_country_groups}
\alias{wb_country_groups}
\title{World Bank country groupings.}
\format{
A tibble with 2,066 rows and 4 variables:
\describe{
  \item{group_code}{World Bank three letter code for the country grouping.}
  \item{group_name}{World Bank country grouping.}
  \item{iso3c}{The iso3c three letter country code}
  \item{country}{The country or region}
  ...
}
}
\source{
\url{http://databank.worldbank.org/data/download/site-content/CLASS.xls}
}
\usage{
wb_country_groups
}
\description{
A dataset in tidy format containing the constituent countries of all
World Bank country groupings.  These groupings are related to geographical
regions, income groups, demographic groups, and other characteristics. A
list of all of the groups in available in the data object
`wb_country_groups_list`
}
\keyword{datasets}
