% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.r
\docType{data}
\name{wb_country_groups_list}
\alias{wb_country_groups_list}
\title{World Bank country groupings.}
\format{
A tibble with 46 rows and 2 variables:
\describe{
  \item{group_code}{World Bank three letter code for the country grouping.}
  \item{group_name}{World Bank country grouping.}
  ...
}
}
\source{
\url{http://databank.worldbank.org/data/download/site-content/CLASS.xls}
}
\usage{
wb_country_groups_list
}
\description{
A dataset in tidy format containing the 46 unique country groupings.  They
are related to region, income groups, demographic groups, and other
features.  The constituent countries of these groups can be found in d
`wb_country_groups``
}
\keyword{datasets}
