% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_color_palette_functions.R
\name{wb_pal}
\alias{wb_pal}
\title{Return function to interpolate a wb color palette}
\usage{
wb_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette wb_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a wb color palette
}
