% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_palettes.R
\docType{data}
\name{wb_palettes}
\alias{wb_palettes}
\title{World Bank Color Palettes}
\format{A character vector of color names and hex values
\describe{
  \item{main}{4 colors. The main blue/black/white palette}
  \item{blue green}{5 colors}
  \item{qualitative 1}{7 colors}
  \item{sequential 1}{7 colors}
  \item{blue green}{4 colors}
  \item{bright warm}{4 colors}
  \item{bright cool}{4 colors}
  \item{bright}{8 colors}
  \item{neutral warm}{3 colors}
  \item{neutral cool}{3 colors}
  \item{neutral}{6 colors}
  \item{warm}{7 colors}
  \item{cool}{7 colors}
  ...
}}
\source{
Colors come from the World Bank's official \href{http://pubdocs.worldbank.org/en/546361452012867685/wbg-branding-visual-identity-guideline.pdf}{branding and visual identity guidelines} color names come \href{http://chir.ag/projects/name-that-color/#6195ED}{this website}. I didn't make them up.  I swear.
}
\usage{
wb_palettes
}
\description{
Color palettes based on the World Bank's official branding and visual identity guidelines.
}
\keyword{datasets}
